pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mina-repl-core"
version = "0.1.0"
description = "Reusable REPL-first skill package for multi-turn AI agents in Python"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "OpenAI ChatGPT", email = "noreply@example.com"}
]
dependencies = [
  "prompt_toolkit>=3.0.50",
  "rich>=14.0.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
ssh = ["asyncssh>=2.21.0"]
textual = ["textual>=0.70.0"]
cmd2 = ["cmd2>=2.5.0"]
dev = ["pytest>=8.0.0"]

[project.scripts]
mina-repl-core = "mina_repl_core.demo:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mina_repl_core = ["data/*.yaml"]
